+++ src/components/layouts/Terms/index.ts
@@ -1,2 +1,3 @@
 export * from './privacy-policy'
 export * from './terms-of-service'
+export * from './tab-content'
